feat: Article과 Comment에 isMine 추가, delete에 검증로직 추가(#64) (#66)
* refactor : ArticleResponse에 isMine 추가

* refactor : ChildCommentResponse에 isMine 추가

* refactor : @Data 제거

* refactor : ParentComment에 isMine 추가

* feat : ArticleController delete함수에 member검증 로직 추가

* feat : ChildController delete함수에 member검증 로직 추가

* feat : ParentController delete함수에 member검증 로직 추가

* refactor: articleService의 findArticles 리턴 객체를 Page로 변경

* refactor:delete 검증로직 오류 수정

member 확인 후 게시글을 삭제하도록 변경

+++ hicc-reboot/src/main/java/hiccreboot/backend/service/ParentCommentService.java
@@ -9,7 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import hiccreboot.backend.common.dto.BaseResponse;
 import hiccreboot.backend.common.dto.DataResponse;
 import hiccreboot.backend.common.dto.ParentComment.ParentCommentResponse;
+import hiccreboot.backend.common.exception.AccessForbiddenException;
 import hiccreboot.backend.common.exception.ArticleNotFoundException;
+import hiccreboot.backend.common.exception.CommentNotFoundException;
 import hiccreboot.backend.common.exception.MemberNotFoundException;
 import hiccreboot.backend.domain.Article;
 import hiccreboot.backend.domain.Member;
@@ -31,7 +33,9 @@ public class ParentCommentService {
 		return parentCommentRepository.findAllByArticle_Id(articleId);
 	}
 
-	public BaseResponse makeParentComments(Long articleId) {
+	public BaseResponse makeParentComments(Long articleId, String studentNumber) {
+		Member member = memberRepository.findByStudentNumber(studentNumber)
+			.orElseThrow(() -> MemberNotFoundException.EXCEPTION);
 
 		List<ParentComment> parentComments = findParentComments(articleId);
 
@@ -41,6 +45,7 @@ public class ParentCommentService {
 				articleId,
 				parentComment.getId(),
 				parentComment.getMember().getName(),
+				parentComment.getMember() == member,
 				parentComment.getDate(),
 				parentComment.getContent())));
 		return DataResponse.ok(parentComments);
@@ -58,7 +63,16 @@ public class ParentCommentService {
 	}
 
 	@Transactional
-	public void deleteParentComment(Long id) {
+	public void deleteParentComment(Long id, String studentNumber) {
+		Member member = memberRepository.findByStudentNumber(studentNumber)
+			.orElseThrow(() -> MemberNotFoundException.EXCEPTION);
+		ParentComment parentComment = parentCommentRepository.findById(id)
+			.orElseThrow(() -> CommentNotFoundException.EXCEPTION);
+
+		if (parentComment.getMember() != member) {
+			throw AccessForbiddenException.EXCEPTION;
+		}
+
 		parentCommentRepository.deleteById(id);
 	}
 }
